ai_dungeon/requirements.txt
pyxel==2.0.5
google-generativeai==0.3.2
python-dotenv==1.0.0
requests==2.31.0  # For downloading sprite assets